move norm to signal.py
